removed IEncoder interface. created BaseEncoder abstract class. keeps base types & methods. Iherited actual encoders.
