--- Encoder/IEncoder.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Encoder
-{
-    public interface IEncoder
-    {
-        void Encode(byte[] source);
-    }
-}
